HULK SMASH PUNY ADAPTERS
